v1.2.9
- Updated project with an improved file structure
- Added 2 new commands: -v, -g
- Now displays version information when using the -v flag
- Opens an experimental gui when using the -g flag. This is a heavy WIP right now with little functionality
